requirements.txt
absl-py==2.2.2
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cors==1.0.1
cycler==0.12.1
fastapi==0.104.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
future==1.0.0
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==1.26.0
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
Pillow==10.1.0
protobuf==5.29.4
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
requests-file==2.1.0
rich==14.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tldextract==5.1.3
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.23.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
